+++ src/frontend/apps/impress/src/features/left-panel/components/LefPanelTargetFilters.tsx
@@ -7,14 +7,12 @@ import { Box, Icon, StyledLink, Text } from '@/components';
 import { useCunninghamTheme } from '@/cunningham';
 import { DocDefaultFilter } from '@/docs/doc-management';
 import { useLeftPanelStore } from '@/features/left-panel';
-import { useResponsiveStore } from '@/stores';
 
 export const LeftPanelTargetFilters = () => {
   const { t } = useTranslation();
   const pathname = usePathname();
   const searchParams = useSearchParams();
 
-  const { isDesktop } = useResponsiveStore();
   const { closePanel } = useLeftPanelStore();
   const { colorsTokens, spacingsTokens } = useCunninghamTheme();
 
@@ -52,9 +50,7 @@ export const LeftPanelTargetFilters = () => {
   };
 
   const handleFilterClick = () => {
-    if (!isDesktop) {
-      closePanel();
-    }
+    closePanel({ type: 'mobile' });
   };
 
   return (