+++ src/frontend/apps/impress/src/features/left-panel/components/index.tsx
@@ -1,3 +1,4 @@
 export * from './LeftPanel';
 export * from './LeftPanelCollapseButton';
+export * from './LeftPanelToggleMobile';
 export * from './ResizableLeftPanel';